+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   (`osmium::memory::Buffer::select()`), over specific subitems
   (`osmium::OSMObject::subitems()`), and for iterating over all rings of
   an area (`osmium::Areas::outer_rings(`), `inner_rings()`).
+- Debug output optionally prints CRC32 when "add_crc32" file option is set.
 
 ### Changed